voice-ai/requirements.txt
# Echo Desk Voice AI - run from project root with: source venv/bin/activate && pip install -r voice-ai/requirements.txt
faster-whisper>=1.0.0
piper-tts>=1.4.0
ollama>=0.1.0
sounddevice>=0.4.0
webrtcvad>=2.0.10
numpy>=1.24.0
scipy>=1.10.0
websockets>=12.0
httpx>=0.25.0
